requirements.txt
numpy>=1.24.0
scipy>=1.10.0
nibabel>=5.0.0
SimpleITK>=2.3.0
matplotlib>=3.7.0
pyyaml>=6.0
